change logs CLI command to be a cmd group, with a default cmd for backwards comliance.
